feat: Comments addressed
