+++ __tests__/data/satismeter_source.json
@@ -228,7 +228,7 @@
           }
     },
     {
-      "description": " Neither reponse.user.id or response.user..userId is provided in payload",
+      "description": " Neither reponse.user.id or response.user.userId is provided in payload then mapping response.id to anonymousId",
       "input":{
           "response": {
             "id": "63767499ca6fb1b7c988d5bb",
@@ -345,7 +345,7 @@
         },
       "output":{
           "event": "Survey completed",
-          "anonymousId":"6a6c19fea4a3676970167ce51f39e6ee",
+          "anonymousId":"63767499ca6fb1b7c988d5bb",
           "context": {
             "library": {
               "name": "unknown",
